[REF] mail:tests: move all refactoring test files in /new folder
