+++ addons/mail/static/tests/new/activity/activity_tests.js
@@ -1,5 +1,6 @@
 /** @odoo-module **/
 
+import { start, startServer } from "@mail/../tests/helpers/test_utils";
 import { Activity } from "@mail/new/activity/activity";
 import { click, getFixture, mount } from "@web/../tests/helpers/utils";
 import { makeTestEnv, TestServer } from "../helpers/helpers";
@@ -46,5 +47,35 @@ QUnit.module("mail", (hooks) => {
         });
         await click(document.querySelector(".o-mail-activity-unlink-button"));
         assert.verifySteps(["/web/dataset/call_kw/mail.activity/unlink"]);
+    });
+
+    QUnit.test("activity upload document is available", async function (assert) {
+        assert.expect(3);
+
+        const pyEnv = await startServer();
+        const resPartnerId1 = pyEnv["res.partner"].create({});
+        const uploadActivityTypeId = pyEnv["mail.activity.type"].search([
+            ["name", "=", "Upload Document"],
+        ])[0];
+        pyEnv["mail.activity"].create({
+            activity_category: "upload_file",
+            activity_type_id: uploadActivityTypeId,
+            can_write: true,
+            res_id: resPartnerId1,
+            res_model: "res.partner",
+        });
+        const { openView } = await start();
+        await openView({
+            res_id: resPartnerId1,
+            res_model: "res.partner",
+            views: [[false, "form"]],
+        });
+        assert.containsOnce(
+            document.body,
+            ".o-mail-activity-name:contains('Upload Document')",
+            "Should have upload document activity"
+        );
+        assert.containsOnce(document.body, ".fa-upload", "Should have activity upload button");
+        assert.containsOnce(document.body, ".o_input_file", "Should have a file uploader");
     });
 });
