+++ addons/mail/static/tests/new/attachment/file_uploader_tests.js
