+++ addons/mail/static/tests/new/views/chatter_prev_tests.js
