+++ addons/mail/static/tests/new/views/follow_button_tests.js
