+++ addons/mail/static/tests/new/views/follower_list_menu_tests.js
