+++ addons/mail/static/tests/new/views/follower_subtype_tests.js
