+++ addons/mail/static/tests/new/views/follower_tests.js
