+++ addons/mail/static/tests/qunit_suite_tests/components/activity_tests.js
@@ -750,36 +750,6 @@ QUnit.module("mail", {}, function () {
             assert.verifySteps(["activity_send_mail"], "should have called activity_send_mail rpc");
         });
 
-        QUnit.test("activity upload document is available", async function (assert) {
-            assert.expect(3);
-
-            const pyEnv = await startServer();
-            const resPartnerId1 = pyEnv["res.partner"].create({});
-            const uploadActivityTypeId = pyEnv["mail.activity.type"].search([
-                ["name", "=", "Upload Document"],
-            ])[0];
-            pyEnv["mail.activity"].create({
-                activity_category: "upload_file",
-                activity_type_id: uploadActivityTypeId,
-                can_write: true,
-                res_id: resPartnerId1,
-                res_model: "res.partner",
-            });
-            const { openView } = await start();
-            await openView({
-                res_id: resPartnerId1,
-                res_model: "res.partner",
-                views: [[false, "form"]],
-            });
-            assert.containsOnce(
-                document.body,
-                ".o-mail-activity-name:contains('Upload Document')",
-                "Should have upload document activity"
-            );
-            assert.containsOnce(document.body, ".fa-upload", "Should have activity upload button");
-            assert.containsOnce(document.body, ".o_input_file", "Should have a file uploader");
-        });
-
         QUnit.skipRefactoring("activity click on mark as done", async function (assert) {
             assert.expect(4);
